+++ CHANGELOG.md
@@ -31,6 +31,10 @@
 * Handle symbols in `Lint/Void`. ([@bbatsov][])
 * [#1695](https://github.com/bbatsov/rubocop/pull/1695): Fix bug with `--auto-gen-config` and `SpaceInsideBlockBraces`. ([@meganemura][])
 
+### Changes
+
+* `UnneededPercentX` renamed to `CommandLiteral`. ([@bquorning][])
+
 ## 0.29.1 (13/02/2015)
 
 ### Bugs fixed